*feat, (fac2tpu), se cambio la forma de obtener la empresa y lugar de expedicion, ahora se maneja mediante la clase "Formatea_Nempresas" con su respectiva funcion requerida. *fix, (daoxmltpu), Se corrigio la entrada de un valor, ya que se la variable es double y se estaba formateando como flotante. *feat, (generarXML40tpu), Se cambio la manera de accesar a los metodos de encriptacion y creacion del xml con la clase de "Formateo_Nempresas"

+++ src/DAO/daoxmltpu.java
@@ -91,7 +91,7 @@ public class daoxmltpu {
                 //Obtener datos de arreglo
                 double importe = f.getArr().get(i).getImporta();
                 double base = f.getArr().get(i).getBase();
-                double unitario = Float.parseFloat(formateador.format(f.getArr().get(i).getPrecio()));
+                double unitario = Double.parseDouble(formateador.format(f.getArr().get(i).getPrecio()));
                 //Cantidad flotante
                 double cantidad = f.getArr().get(i).getCantidadfloat();
                 double descuento = f.getArr().get(i).getDescuento();
